% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual_plots.r
\name{accrual_plot_cum}
\alias{accrual_plot_cum}
\title{accrual_plot_cum}
\usage{
accrual_plot_cum(
  accrual_df,
  overall = TRUE,
  name_overall = "Overall",
  start_date = NA,
  current_date = NA,
  ylim = NA,
  xlim = NA,
  ylab = "Recruited patients",
  xlabn = 5,
  xlabminn = xlabn\%/\%2,
  xlabformat = "\%d\%b\%Y",
  xlabpos = NA,
  xlabsrt = 45,
  xlabadj = c(1, 1),
  xlabcex = 1,
  col = rep(1:8, 5),
  lty = rep(1:5, each = 8),
  legend.list = NULL,
  ...
)
}
\arguments{
\item{accrual_df}{accrual data frame produced by accrual_create_df potentially with by option (i.e. as a list)}

\item{overall}{indicates that accrual_df contains a summary with all sites (only if by is not NA)}

\item{name_overall}{name of the summary with all sites (if by is not NA and overall==TRUE)}

\item{start_date}{start_date: date when recruitment started, single character or date,
or "common" if the same date should be used for all sites,
  if not given the first enrollment date is used as start_date}

\item{current_date}{date of the data export or database freeze, single character or date
or "common" if the same date should be used for all sites,
  if not given the latest enrollment date is used for each site}

\item{ylim}{limits for y-axis}

\item{xlim}{limits for x-axis}

\item{ylab}{y-axis label}

\item{xlabn}{integer giving the desired number of intervals for the xlabel, default=5}

\item{xlabminn}{nonnegative integer giving the minimal number of intervals}

\item{xlabformat}{format of date on x-axis}

\item{xlabpos}{position of the x-label}

\item{xlabsrt}{rotation of x-axis labels in degrees}

\item{xlabadj}{adjustment of x-label, numeric vector with length 1 or 2 for different adjustment in x- and y-direction}

\item{xlabcex}{size of x-axis label}

\item{col}{color for line(s) in plot}

\item{lty}{line types in plot}

\item{legend.list}{named list with options passed to legend()}

\item{...}{further options passed to plot() and axis()}
}
\value{
A plot of the cumulative accrual, optionally by site.
}
\description{
Plot of cumulative recruitment based on accrual data frame produced by accrual_create_df
}
\examples{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
accrual_df<-accrual_create_df(enrollment_dates)
accrual_plot_cum(accrual_df)
accrual_plot_cum(accrual_df,cex.lab=1.2,cex.axis=1.1,xlabcex=1.1)

#several sites
set.seed(1)
centers<-sample(c("Site 1","Site 2","Site 3"),length(enrollment_dates),replace=TRUE)
accrual_df<-accrual_create_df(enrollment_dates,by=centers)
accrual_plot_cum(accrual_df)

#assuming a common start and current date
accrual_plot_cum(accrual_df,start_date="common",current_date="common")

#plot and legend options
accrual_plot_cum(accrual_df,start_date="common",current_date="common",
     col=c("red",rep(1,3)),lty=c(1,1:3),cex.lab=1.2,cex.axis=1.1,xlabcex=1.1)
accrual_plot_cum(accrual_df,legend.list=list(ncol=2,bty=TRUE,cex=0.8))

#without overall
accrual_df<-accrual_create_df(enrollment_dates,by=centers,overall=FALSE)
accrual_plot_cum(accrual_df,start_date="common",current_date="common",overall=FALSE)

}
