% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual_plots.r
\name{accrual_table}
\alias{accrual_table}
\title{accrual_table}
\usage{
accrual_table(
  accrual_df,
  overall = TRUE,
  name_overall = "Overall",
  pos_overall = c("last", "first"),
  start_date = NA,
  format_start_date = "\%d\%b\%Y",
  current_date = NA,
  format_current_date = "\%d\%b\%Y",
  unit = c("month", "year", "week", "day"),
  format_table_date = "\%d\%b\%Y",
  format_time = "\%1.0f",
  format_rrate = "\%1.2f",
  header = TRUE
)
}
\arguments{
\item{accrual_df}{accrual data frame produced by accrual_create_df potentially with by option (i.e. as a list)}

\item{overall}{logical, indicates that accrual_df contains a summary with all sites (only if by is not NA)}

\item{name_overall}{name of the summary with all sites (if by is not NA and overall==TRUE)}

\item{pos_overall}{overall in last or first row (if by is not NA and overall==TRUE)}

\item{start_date}{start_date: date when recruitment started, single character or date,}

\item{format_start_date}{format of the start date, ignored if start_date is a date}

\item{current_date}{date of the data export or database freeze, single character or date}

\item{format_current_date}{format of the current date, ignored if current_date is a date}

\item{unit}{time unit for time recruiting and the rate, any of "month","year","week","day"}

\item{format_table_date}{format of start date in table}

\item{format_time}{format of time recruiting in table}

\item{format_rrate}{format of recruitment rate in table}

\item{header}{header: include header, TRUE, NULL or character vector of length 5}
}
\value{
Data frame with name of the site, accrual start date, time accruing,
 number of patients accrued and accrual rate.
}
\description{
Table of recruitment overview by site,
including name of site, start date, time recruiting, number of patients recruited and
rate of recruitment
}
\examples{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
centers<-sample(c("Site 1","Site 2","Site 3"),length(enrollment_dates),replace=TRUE)
accrual_df<-accrual_create_df(enrollment_dates,by=centers)
accrual_table(accrual_df)

#format
accrual_table(accrual_df,format_time="\%1.1f",format_rrate="\%1.1f")

#unit
accrual_table(accrual_df,unit="day")

#common start and current dates
accrual_table(accrual_df,unit="day",start_date="common",current_date="common")
accrual_table(accrual_df,unit="day",start_date=as.Date("2017-12-31"),
    current_date=as.Date("2018-03-01"))


}
