% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual_time_unit.R
\name{accrual_time_unit}
\alias{accrual_time_unit}
\title{accrual_time_unit}
\usage{
accrual_time_unit(accrual_df, unit = c("month", "year", "week", "day"))
}
\arguments{
\item{accrual_df}{accrual data frame produced by accrual_create_df with by=NA}

\item{unit}{time unit for which the bars should be plotted, any of "month","year","week","day"}
}
\value{
Data frame with the number of patients accrued for each time unit
(given by day and/or week and/or month and year).
}
\description{
Generates summary of recruitment per time unit
}
\examples{
\donttest{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
accrual_df<-accrual_create_df(enrollment_dates)
accrualPlot:::accrual_time_unit(accrual_df,"week")
accrualPlot:::accrual_time_unit(accrual_df,"day")
}

}
