% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accrualdemo}
\alias{accrualdemo}
\title{Demonstration data set}
\format{
A data frame with two variables: \code{date}, and \code{site}.
}
\usage{
accrualdemo
}
\description{
Simulated recruitment data from three sites. Each row represents one participant. Sites one and two started on 2020-07-01, site three on 2020-09-01.
}
\keyword{datasets}
