% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.accrual_list}
\alias{as.data.frame.accrual_list}
\title{as.data.frame method for accural_list objects}
\usage{
\method{as.data.frame}{accrual_list}(x)
}
\arguments{
\item{x}{accrual_list}
}
\description{
as.data.frame method for accural_list objects
}
\examples{
data(accrualdemo)
x <- accrual_create_df(accrualdemo$date, accrualdemo$site)
as.data.frame(x)
x <- accrual_create_df(accrualdemo$date)

}
