% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.accrual_df}
\alias{plot.accrual_df}
\title{Plot method for accrual_dfs (as created by accrual_create_df)}
\usage{
\method{plot}{accrual_df}(x, which = "cum", ...)
}
\arguments{
\item{x}{accrual_df object}

\item{which}{one of \code{"cumulative"}, \code{"absolute"} or \code{"predict"}.
Abbreviations are allowed.}

\item{...}{options passed to other functions}
}
\value{
a plot
}
\description{
Plot method for accrual_dfs (as created by accrual_create_df)
}
