% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual_create_df.R
\name{accrual_create_df}
\alias{accrual_create_df}
\title{accrual_create_df}
\usage{
accrual_create_df(
  enrollment_dates,
  start_date = NA,
  current_date = NA,
  force_start0 = TRUE,
  by = NA,
  overall = TRUE,
  name_overall = "Overall"
)
}
\arguments{
\item{enrollment_dates}{dates on which patients are enrolled as date or character vector}

\item{start_date}{date when recruitment started, single character or date, if not given the first enrollment date is used}

\item{current_date}{date of the data export or database freeze, single character or date, if not given the latest enrollment date is used}

\item{force_start0}{adds an extra 0 line to the accrual data frame in cases where a start date is given and
corresponds to the earliest enrollment date}

\item{by}{vector with centers, has to have the same length as enrollment dates,
generates a list with accrual data frames for each site}

\item{overall}{indicates that accrual_df contains a summary with all sites (only if by is not NA)}

\item{name_overall}{name of the summary with all sites (if by is not NA and overall==TRUE)}

\item{format_enrollment_dates}{single character identifying the format of the dates,
ignored if enrollment_dates are dates}
}
\value{
Returns a data frame (or a list of data frames if by is not NA)
with three columns "Date", "Freq" and "Cumulative" with each
date with an accural and the absolute and cumulative number of patients accrued.
}
\description{
Creates a data frame that contains the absolute and cumululative number of patients
recruited at each date from a vector with enrollment dates.
}
\examples{
\donttest{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
accrual_create_df(enrollment_dates)
# different start and current date
accrual_create_df(enrollment_dates,start_date=as.Date("2017-12-01"),
current_date=as.Date("2018-03-01"))
#by site
set.seed(2020)
centers<-sample(c("Site 1","Site 2","Site 3"),length(enrollment_dates),replace=TRUE)
accrual_create_df(enrollment_dates,by=centers)
}
}
