% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual_create_df.R
\name{accrual_create_df}
\alias{accrual_create_df}
\title{accrual_create_df}
\usage{
accrual_create_df(
  enrollment_dates,
  start_date = "site",
  current_date = "common",
  force_start0 = TRUE,
  by = NA,
  overall = TRUE,
  name_overall = "Overall",
  pos_overall = c("last", "first")
)
}
\arguments{
\item{enrollment_dates}{dates on which patients are enrolled, date vector}

\item{start_date}{date when recruitment started. Single date (used for all sites in by),
date vector (with the same length the number of distinct sites in by),
"common" (first date overall) or "site" (first date for each site, default).}

\item{current_date}{date of the data export or database freeze.
   Single date, date vector (with the same length the number of distinct sites in by),
"common" (last date overall, default) or "site" (first date for each site).}

\item{force_start0}{logical, adds an extra 0 line to the accrual data frame in cases
where a start date is given and corresponds to the earliest enrollment date.}

\item{by}{vector with sites, has to have the same length as enrollment dates,
generates a list with accrual data frames for each site}

\item{overall}{logical indicates that accrual_df contains a summary with all sites (only if by is not NA)}

\item{name_overall}{name of the summary with all sites (if by is not NA and overall==TRUE)}

\item{pos_overall}{overall as last or first element of the list (if by is not NA and overall==TRUE)}
}
\value{
Returns a data frame (or a list of data frames if by is not NA)
with three columns "Date", "Freq" and "Cumulative" with each
date with an accrual and the absolute and cumulative number of patients accrued.
}
\description{
Creates a data frame that contains the absolute and cumululative number of patients
recruited at each date from a vector with enrollment dates.
}
\examples{
\donttest{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
accrual_create_df(enrollment_dates)
# different start and current date
accrual_create_df(enrollment_dates,start_date=as.Date("2017-12-01"),
current_date=as.Date("2018-03-01"))
#by site
set.seed(2020)
centers<-sample(c("Site 1","Site 2","Site 3"),length(enrollment_dates),replace=TRUE)
accrual_create_df(enrollment_dates,by=centers)
}
}
