% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual_plots.r
\name{accrual_linear_model}
\alias{accrual_linear_model}
\title{accrual_linear_model}
\usage{
accrual_linear_model(
  accrual_df,
  start_date = NA,
  format_start_date = "\%d\%b\%Y",
  current_date = NA,
  format_current_date = "\%d\%b\%Y",
  fill_up = c("yes", "no"),
  wfun = function(x) seq(1/nrow(x), 1, by = 1/nrow(x))
)
}
\arguments{
\item{accrual_df}{accrual data frame produced by accrual_create_df with by=NA}

\item{start_date}{date when recruitment started, single character or date,}

\item{format_start_date}{format of the start date, ignored if start_date is a date}

\item{current_date}{date of the data export or database freeze, single character or date,}

\item{format_current_date}{format of the current date, ignored if current_date is a date}

\item{fill_up}{whether to fill up days where no recruitment was observed,}

\item{wfun}{function to calculate the weights based on the accrual data frame, default is}
}
\value{
A lm object of a weigthed linear regression of cumulative accrual on dates.
}
\description{
Creates weighted linear regression model based on an accrual data frame producd by accrual_create_df.
}
\examples{
\donttest{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
accrual_df<-accrual_create_df(enrollment_dates)
accrual_linear_model(accrual_df)

#unweighted
accrual_linear_model(accrual_df, wfun=function(x) rep(1,nrow(x)))

#different start and current date
accrual_df<-accrual_create_df(enrollment_dates)
accrual_linear_model(accrual_df,start_date=as.Date("2017-12-01"),
    current_date=as.Date("2018-03-01"))
#or
accrual_df<-accrual_create_df(enrollment_dates,start_date=as.Date("2017-12-01"),
    current_date=as.Date("2018-03-01"))
accrual_linear_model(accrual_df)
}
}
