% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual_plots.r
\name{accrual_plot_abs}
\alias{accrual_plot_abs}
\title{accrual_plot_abs}
\usage{
accrual_plot_abs(
  accrual_df,
  unit = c("month", "year", "week", "day"),
  target = NA,
  start_date = NA,
  format_start_date = "\%d\%b\%Y",
  current_date = NA,
  format_current_date = "\%d\%b\%Y",
  ylim = NA,
  xlim = NA,
  ylab = "Recruited patients",
  xlabformat = "\%b \%Y",
  xlabsel = NA,
  xlabpos = NA,
  xlabsrt = 45,
  xlabadj = c(1, 1),
  xlabcex = 1,
  col = "grey",
  ...
)
}
\arguments{
\item{accrual_df}{accrual data frame produced by accrual_create_df with by=NA}

\item{unit}{time unit for which the bars should be plotted, any of "month","year","week","day"}

\item{target}{adds horizontal line for target recruitment per time unit}

\item{start_date}{start_date: date when recruitment started, single character or date,}

\item{format_start_date}{format of the start date, ignored if start_date is a date}

\item{current_date}{date of the data export or database freeze, single character or date}

\item{format_current_date}{format of the current date, ignored if current_date is a date}

\item{ylim}{limits for y-axis}

\item{xlim}{limits for x-axis, in barplot units}

\item{ylab}{y-axis label}

\item{xlabformat}{format of date on x-axis}

\item{xlabsel}{selection of x-labels if not all should be shown,}

\item{xlabpos}{position of the x-label}

\item{xlabsrt}{rotation of x-axis labels in degrees}

\item{xlabadj}{adjustment of x-label, numeric vector with length 1 or 2 for different adjustment
in x- and y-direction}

\item{xlabcex}{size of x-axis label}

\item{col}{colors of bars in barplot}

\item{...}{further arguments passed to barplot() and axis()}
}
\value{
Barplot of absolute recruitment by time unit.
}
\description{
Plot of absolute recruitment by time unit
}
\examples{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
accrual_df<-accrual_create_df(enrollment_dates)
accrual_plot_abs(accrual_df,unit="week",xlabformat="\%d\%b\%Y")

#time unit
accrual_plot_abs(accrual_df,unit="day",xlabformat="\%d\%b\%Y")

#include target
accrual_plot_abs(accrual_df,unit="week",xlabformat="\%d\%b\%Y",target=10)

#different start and current dates
accrual_plot_abs(accrual_df,unit="week",xlabformat="\%d\%b\%Y",target=10,
   start_date=as.Date("2017-12-01"),current_date=as.Date("2018-03-01"))

#further plot options
accrual_plot_abs(accrual_df,unit="week",ylab="No of recruited patients",
   xlabformat="\%Y-\%m-\%d",xlabsrt=30,xlabpos=-0.8,xlabadj=c(1,0.5),
   col="pink",tck=-0.03,mgp=c(3,1.2,0))


}
