% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual_predict.R
\name{accrual_predict}
\alias{accrual_predict}
\title{accrual_predict}
\usage{
accrual_predict(accrual_df, accrual_fit, target)
}
\arguments{
\item{accrual_df}{accrual data frame produced by \code{accrual_create_df} (optionally with by option as a list)}

\item{accrual_fit}{linear model produced by accrual_linear_model, can be a list with the same length as accrual_df}

\item{target}{target sample size, either a single number or a named vector with the same length as accrual_df,
for the latter, center-specific predictions are shown.}
}
\value{
Returns the predicted end date or a list of the predicted end dates
}
\description{
accrual_predict
}
\details{
Prediction of end date based on an accrual data frame produced by \code{accrual_create_df},
a fitted regression model produced by accrual_linear_model and a target sample size.
}
\examples{
\donttest{
data(accrualdemo)
accrual_df<-accrual_create_df(accrualdemo$date)
accrual_model<-accrual_linear_model(accrual_df)
accrual_predict(accrual_df,accrual_model,target=300)

#different start and current date
accrual_df<-accrual_create_df(accrualdemo$date,start_date=as.Date("2020-07-09"),
    current_date=as.Date("2020-10-15"))
accrual_model<-accrual_linear_model(accrual_df)
accrual_predict(accrual_df,accrual_model,target=300)

 #accrual_df with by option
 accrual_df<-accrual_create_df(accrualdemo$date,by=accrualdemo$site)
accrual_model<-accrual_linear_model(accrual_df)
accrual_predict(accrual_df,accrual_model,
target=c("Site 1"=100,"Site 2"=100,"Site 3"=100,"Overall"=350))
}
}
