% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual_predict.R
\name{accrual_predict}
\alias{accrual_predict}
\title{accrual_predict}
\usage{
accrual_predict(
  accrual_df,
  accrual_fit,
  target,
  current_date = NA,
  format_current_date = "\%d\%b\%Y"
)
}
\arguments{
\item{accrual_df}{accrual data frame produced by accrual_create_df with by=NA}

\item{accrual_fit}{linear model produced by accrual_linear_model}

\item{target}{target sample size, single number}

\item{current_date}{date of the data export or database freeze, single character or date,}

\item{format_current_date}{format of the current date, ignored if current_date is a date}
}
\value{
A single date, the predicted end date.
}
\description{
Prediction of end date based on an accrual data frame produced by accrual_create_df,
a fitted regression model produced by accrual_linear_model and a target sample size.
}
\examples{
\donttest{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
accrual_df<-accrual_create_df(enrollment_dates)
accrual_model<-accrual_linear_model(accrual_df)
accrualPlot:::accrual_predict(accrual_df,accrual_model,target=100)

#different start and current date
accrual_df<-accrual_create_df(enrollment_dates,start_date=as.Date("2017-12-01"),
    current_date=as.Date("2018-03-01"))
accrual_model<-accrual_linear_model(accrual_df)
accrual_plot_predict(accrual_df,accrual_model,target=100)
#or
accrual_df<-accrual_create_df(enrollment_dates)
accrual_model<-accrual_linear_model(accrual_df,start_date=as.Date("2017-12-01"),
    current_date=as.Date("2018-03-01"))
accrualPlot:::accrual_predict(accrual_df,accrual_model,target=100,
    current_date=as.Date("2018-03-01"))
}
}
