% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual_time_unit.R
\name{accrual_time_unit}
\alias{accrual_time_unit}
\title{accrual_time_unit}
\usage{
accrual_time_unit(
  accrual_df,
  unit = c("month", "year", "week", "day"),
  start_date = NA,
  format_start_date = "\%d\%b\%Y",
  current_date = NA,
  format_current_date = "\%d\%b\%Y"
)
}
\arguments{
\item{accrual_df}{accrual data frame produced by accrual_create_df with by=NA}

\item{unit}{time unit for which the bars should be plotted, any of "month","year","week","day"}

\item{start_date}{start_date: date when recruitment started, single character or date,}

\item{format_start_date}{format of the start date, ignored if start_date is a date}

\item{current_date}{date of the data export or database freeze, single character or date}

\item{format_current_date}{format of the current date, ignored if current_date is a date}
}
\value{
Data frame with the number of patients accrued for each time unit
(given by day and/or week and/or month and year).
}
\description{
Generates summary of recruitment per time unit
}
\examples{
\donttest{
set.seed(2020)
enrollment_dates <- as.Date("2018-01-01") + sort(sample(1:30, 50, replace=TRUE))
accrual_df<-accrual_create_df(enrollment_dates)
accrualPlot:::accrual_time_unit(accrual_df,"week")
accrualPlot:::accrual_time_unit(accrual_df,"day")
accrualPlot:::accrual_time_unit(accrual_df,"week",
    start_date=as.Date("2017-12-01"),
    current_date=as.Date("2018-03-01"))
}
}
